refactoring buttons part 1
